only notes in event or task

+++ models/Branch.js
@@ -311,13 +311,15 @@ const deleteTaskOrCategoryRecursive = async (type, id)=>{
             }
         )
     );
-    result.children.notes = await Promise.all(
-        notes.map(async noteId=>{
-                const noteIdString = noteId.toString();
-                await Note.findOneAndRemove({_id:noteIdString});
-            }
-        )
-    );
+    if(notes) {
+        result.children.notes = await Promise.all(
+            notes.map(async noteId => {
+                    const noteIdString = noteId.toString();
+                    await Note.findOneAndRemove({_id: noteIdString});
+                }
+            )
+        );
+    }
     return result;
 
 };
